requirements.txt
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.15.0
streamlit>=1.28.0
kaggle>=1.5.16
folium>=0.14.0
streamlit-folium>=0.15.0
python-dotenv>=1.0.0
scipy>=1.11.0
openpyxl>=3.1.0
